requirements.txt
hikari==2.1.0
hikari_lightbulb==2.3.5.post1
pydantic==2.9.2
pymongo==4.10.1
python-dotenv==1.0.1